[stream-restore-nemo] Update patchset on top of latest module.
The module was forked quite a long time ago from PulseAudio. Merge the
latest changes (PulseAudio 8.0) to the module and also some slight
refactoring of the implementation to try to be as closely the upstream
module with our changes on top to see more easily the changed bits
from upstream version.