[mainvolume] Mute media streams when volumes are changing.
When audio route is changing there is a brief moment when volumes are
out of sync, as both routing change and volume update for the new route
are asynchronous operations.

Now when the element handling routing change (here
pulseaudio-policy-enforcement module) starts to do a routing change, it
signals about that and mainvolume then mutes audio, also when mainvolume
receives volume change notification mainvolume adds to that non-sync
state. This way when both are done with the change we know to unmute the
stream volumes and there won't be brief loud or quiet audio from wrong
audio route when changing the routing.